Remove RocketData and StatusDataBlock
